requirements.txt
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
Deprecated==1.2.13
firebase-admin==5.2.0
gcloud==0.18.3
google-api-core==2.8.2
google-api-python-client==2.56.0
google-auth==2.10.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.6.0
google-cloud-secret-manager==2.12.2
google-cloud-storage==2.5.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
grpc-google-iam-v1==0.12.4
grpcio==1.47.0
grpcio-status==1.47.0
httplib2==0.20.4
idna==3.3
jwcrypto==1.3.1
msgpack==1.0.4
oauth2client==4.1.3
proto-plus==1.22.0
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.15.0
pyparsing==3.0.9
Pyrebase4==4.5.0
python-jwt==3.3.3
requests==2.28.1
requests-toolbelt==0.9.1
rsa==4.9
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.11
wrapt==1.14.1
